requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
anthropic==0.7.0